# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JPOTTA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
}

export interface Review {
  id: number;
  authorName: string;
  /** @nullable */
  authorTitle?: string | null;
  content: string;
  rating: number;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface CreateReviewBody {
  authorName: string;
  /** @nullable */
  authorTitle?: string | null;
  content: string;
  rating: number;
  /** @nullable */
  photoUrl?: string | null;
}

export interface UpdateReviewBody {
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorTitle?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  photoUrl?: string | null;
}

export interface NewsArticle {
  id: number;
  title: string;
  summary: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
}

export interface CreateNewsArticleBody {
  title: string;
  summary: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  publishedAt: string;
}

export interface UpdateNewsArticleBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  publishedAt?: string | null;
}

export interface Achievement {
  id: number;
  title: string;
  shortDescription: string;
  longDescription: string;
  year: number;
  /** @nullable */
  mainImageUrl?: string | null;
  additionalImages: string[];
  taggedPlayerIds: number[];
  createdAt: string;
}

export interface Player {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  highestRank?: string | null;
  /** @nullable */
  blade?: string | null;
  /** @nullable */
  bhRubber?: string | null;
  /** @nullable */
  fhRubber?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface AchievementDetail {
  id: number;
  title: string;
  shortDescription: string;
  longDescription: string;
  year: number;
  /** @nullable */
  mainImageUrl?: string | null;
  additionalImages: string[];
  taggedPlayers: Player[];
  createdAt: string;
}

export interface CreateAchievementBody {
  title: string;
  shortDescription: string;
  longDescription: string;
  year: number;
  /** @nullable */
  mainImageUrl?: string | null;
  additionalImages?: string[];
  taggedPlayerIds?: number[];
}

export interface UpdateAchievementBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  mainImageUrl?: string | null;
  additionalImages?: string[];
  taggedPlayerIds?: number[];
}

export interface TagPlayerBody {
  playerId: number;
}

export interface PlayerDetail {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  highestRank?: string | null;
  /** @nullable */
  blade?: string | null;
  /** @nullable */
  bhRubber?: string | null;
  /** @nullable */
  fhRubber?: string | null;
  displayOrder: number;
  achievements: Achievement[];
  createdAt: string;
}

export interface CreatePlayerBody {
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  highestRank?: string | null;
  /** @nullable */
  blade?: string | null;
  /** @nullable */
  bhRubber?: string | null;
  /** @nullable */
  fhRubber?: string | null;
  displayOrder?: number;
}

export interface UpdatePlayerBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  highestRank?: string | null;
  /** @nullable */
  blade?: string | null;
  /** @nullable */
  bhRubber?: string | null;
  /** @nullable */
  fhRubber?: string | null;
  /** @nullable */
  displayOrder?: number | null;
}

export interface CommitteeMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface CreateCommitteeMemberBody {
  name: string;
  role: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  displayOrder?: number;
}

export interface UpdateCommitteeMemberBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  displayOrder?: number | null;
}

export type ListAchievementsParams = {
  /**
   * @nullable
   */
  year?: number | null;
};

export type ListPlayersParams = {
  /**
   * @nullable
   */
  search?: string | null;
};