# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JPOTTA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── News ────────────────────────────────────────────────────────────
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create player (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerBody"
      responses:
        "201":
          description: Created player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player with stats and achievements history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Committee Members ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: integer
        authorName:
          type: string
        authorTitle:
          type: ["string", "null"]
        content:
          type: string
        rating:
          type: integer
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - authorName
        - content
        - rating
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        authorName:
          type: string
        authorTitle:
          type: ["string", "null"]
        content:
          type: string
        rating:
          type: integer
        photoUrl:
          type: ["string", "null"]
      required:
        - authorName
        - content
        - rating
    UpdateReviewBody:
      type: object
      properties:
        authorName:
          type: ["string", "null"]
        authorTitle:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        photoUrl:
          type: ["string", "null"]
...
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortDescription:
          type: string
        longDescription:
          type: string
        year:
          type: integer
        mainImageUrl:
          type: ["string", "null"]
        additionalImages:
          type: array
          items:
            type: string
        taggedPlayerIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - shortDescription
        - longDescription
        - year
        - additionalImages
        - taggedPlayerIds
        - createdAt
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        highestRank:
          type: ["string", "null"]
        blade:
          type: ["string", "null"]
        bhRubber:
          type: ["string", "null"]
        fhRubber:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - displayOrder
        - createdAt
    PlayerDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        highestRank:
          type: ["string", "null"]
        blade:
          type: ["string", "null"]
        bhRubber:
          type: ["string", "null"]
        fhRubber:
          type: ["string", "null"]
        displayOrder:
          type: integer
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        createdAt:
          type: string
      required:
        - id
        - name
        - displayOrder
        - achievements
        - createdAt
    CreatePlayerBody:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        highestRank:
          type: ["string", "null"]
        blade:
          type: ["string", "null"]
        bhRubber:
          type: ["string", "null"]
        fhRubber:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - name
    UpdatePlayerBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        highestRank:
          type: ["string", "null"]
        blade:
          type: ["string", "null"]
        bhRubber:
          type: ["string", "null"]
        fhRubber:
          type: ["string", "null"]
        displayOrder:
          type: ["integer", "null"]
...