# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JPOTTA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  highestRank?: string | null;
  /** @nullable */
  blade?: string | null;
  /** @nullable */
  bhRubber?: string | null;
  /** @nullable */
  fhRubber?: string | null;
  displayOrder: number;
  createdAt: string;
}